Changed legend functionality and display of coords
Due to Ana Verulidze: The legend now allows the user to click a square in the legend to change how clicking the grid works. As a result, the square they clicked will determine how much the states of the squares below change.
Due largely to Ana Verulidze: There is now a checkbox that determines if coordinates or just the number of clicks appears as the text in the cells.